Fix tests for compatibility with pypy
The pypy garbage collector can not be relied upon to be run immediately after
the deletion of the last hard reference to an object. Our tests previous relied
on this implementation detail of CPython, which is not part of the
specification though. Therefore, the garbage collection has to be run manually
to ensure proper execution of the tests in pypy.