Add python3 specific tests for preserving the signature
Sorry for the extra top level folder. It seems nosetests has no real
support for excluding individual tests based on runtime version.

There are the --py3where and --where options that could be used. Though,
with one of these enabled the entire test is run from the specified
subdirectory. Which means that all top level content has to be symlinked
in order to provide a valid runtime without the module installed.