Add coveralls.io support
Coveralls.io can display the coverage of the unit tests.

To generate coverage metrics on Travis, we run the tests with the coverage
package. After a successful build the metrics are pushed to
coveralls.io.

As we supply an empty implementation for a python-3-only function, we don't want
to track the coverage of that code path, as its execution depends on the Python version.