Add pypip.in badges for current version, license and downloads on pypi
pypip.in provides badges for different metrics pulled from pypi, e.g. version,
license, recent downloads, if the project provides eggs or wheels, etc. Those are
used in the README, to provide useful informations to a visitor of the projects
github page.

The build status badge from Travis is also added to the top of the README, so it
is better visible on the github page.

Other badges should also be added at the top for visibility.