Cache the boundevent object with the instance
The boundevent is now cached as attribute of the instance and never gets
garbage collected before its owner object.

The list of event handlers becomes an instance variable of the
boundevent object. This is reasonable since it is (and always was) the
boundevent object's responsibility to manage the event handlers.

Furthermore, this makes the boundevent object more self-contained
(independent of the instance it is attached to), which will make it
usable on its own.