Revert caching of the boundevent instance
The caching of the boundevent instance will have the effect that it
the reference count of the object which it is tied to can never reach
zero. It can be cleaned up only by the garbage collector. This may be
very undesirable (and unexpected) in some circumstances.

Instead, we can use the event_handler argument of boundevent to create
boundevents that are specific to some instance.