Disable class-based connection of instance-specific handlers
The following signatures are removed:

    cls.event.connect(instance, handler)

The following stays valid:

    cls.event.connect(handler)
    inst.event.connect(handler)

The rationale is simplicity: It is easier to remember a single calling
signature with only one argument. In particular the removed signature
might be confusing in terms of parameter order, since the *first*
parameter was optional.

Furthermore, this is easier to handle internally, especially with regard
to upcoming revival of a bound_event class.